Update files to fix segfaulting on server.

+++ src/tinyfile_service.h
@@ -2,6 +2,7 @@
 #define __TINYFILE_SERVICE_H
 
 #include <stdlib.h>
+#include <stdint.h>
 
 #define NUM_THREADS 5
 #define MAX_THREADS 8
@@ -54,8 +55,8 @@ typedef struct HashMap {
     HashNode* buckets[HASH_MAP_SIZE];
 } HashMap;
 
-long fnv1a_hash(long data) {
-    long hash = FNV_32_OFFSET_BASIS;
+uint32_t fnv1a_hash(uint32_t data) {
+    uint32_t hash = FNV_32_OFFSET_BASIS;
     for (int i = 0; i < 4; i++) {
         hash ^= (data & 0xFF);
         hash *= FNV_32_PRIME;
